backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-aiplatform==1.38.0
google-cloud-translate==3.15.0
google-cloud-language==2.13.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
databases==0.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
numpy==1.26.3
tenacity==8.2.3
